Move NDP app to ndp.lua
